+++ Spotifly/Store/EntityConversions.swift
@@ -92,7 +92,6 @@ extension Artist {
         uri = artist.uri
         imageURL = artist.imageURL
         genres = artist.genres
-        followers = artist.followers
         externalUrl = artist.externalUrl
     }
 }
@@ -104,12 +103,8 @@ extension UserProfile {
     init(from profile: UserProfileCodable) {
         id = profile.id
         displayName = profile.displayName ?? profile.id
-        email = profile.email
         imageURL = profile.images?.first.flatMap { URL(string: $0.url) }
         externalUrl = profile.externalUrls?.spotify
-        followers = profile.followers?.total
-        product = profile.product
-        country = profile.country
         uri = profile.uri
     }
 }